turn on eslint rule no-unused-vars
