+++ src/components/Pokedex/type.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import style from './styles.module.css';
 import { Typography } from '@mui/material';
